// src/api/common/CurrencyFormat.ts
// AUTO-GENERATED BY tools/generate_types.py
// DO NOT EDIT MANUALLY

export type CurrencyFormat = {
  iso_code: string;
  example_format: string;
  decimal_digits: number;
  decimal_separator: string;
  symbol_first: boolean;
  group_separator: string;
  currency_symbol: string;
  display_symbol: boolean;
} | null;